Move partition splitting driver code to its own package
