+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/splitting/PartitionSplittingDriver.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.systemtest.drivers.compaction;
+package sleeper.systemtest.drivers.splitting;
 
 import software.amazon.awssdk.services.lambda.LambdaClient;
