+++ java/system-test/system-test-drivers/src/test/java/sleeper/systemtest/drivers/compaction/WaitForPartitionSplittingTest.java
@@ -28,6 +28,7 @@ import sleeper.core.schema.Schema;
 import sleeper.core.schema.type.StringType;
 import sleeper.core.statestore.StateStore;
 import sleeper.core.statestore.inmemory.StateStoreTestBuilder;
+import sleeper.systemtest.drivers.splitting.WaitForPartitionSplitting;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.configuration.properties.InstancePropertiesTestHelper.createTestInstanceProperties;
